changed where interface to require with_index
I'm dropping the code that determines if it the selector takes one or two
params. It is unrealiable as it doesn't work on built-in callables. The
methods that are overloaded to accept one or two parameter funcs will
have have to be told explicitely how which kind of callable is being used.